+++ apps/root/src/pages/transfer/transfer-container/components/transfer-button/index.tsx
@@ -1,16 +1,16 @@
 import React from 'react';
 import { parseUnits } from '@ethersproject/units';
-import useContactListService from '@hooks/useContactListService';
 import { useTransferState } from '@state/transfer/hooks';
 import { FormattedMessage } from 'react-intl';
 import { Button, Typography } from 'ui-library';
 import { PROTOCOL_TOKEN_ADDRESS } from '@common/mocks/tokens';
 import { Token, TokenType } from '@types';
 import useActiveWallet from '@hooks/useActiveWallet';
+import useWalletService from '@hooks/useWalletService';
 
 const TransferButton = () => {
   const { token, amount, recipient } = useTransferState();
-  const contactListService = useContactListService();
+  const walletService = useWalletService();
   const activeWallet = useActiveWallet();
 
   const parsedAmount = parseUnits(amount || '0', token?.decimals);
@@ -22,11 +22,11 @@ const TransferButton = () => {
     }
     try {
       const isProtocolToken = token.address === PROTOCOL_TOKEN_ADDRESS;
-      const parsedToken: Token = { ...token, type: isProtocolToken ? TokenType.BASE : TokenType.ERC20_TOKEN };
+      const parsedToken: Token = { ...token, type: isProtocolToken ? TokenType.NATIVE : TokenType.ERC20_TOKEN };
 
-      await contactListService.transferTokenToContact({
+      await walletService.transferToken({
         from: activeWallet.address,
-        contact: { address: recipient },
+        to: recipient,
         token: parsedToken,
         amount: parsedAmount,
       });
