+++ apps/root/src/services/contactListService.spec.ts
@@ -1,8 +1,5 @@
 import { createMockInstance } from '@common/utils/tests';
-import { Contact, ERC20Contract, ERC721Contract, TokenType, User, UserType } from '@types';
-import { emptyTokenWithAddress } from '@common/utils/currency';
-import { BigNumber } from 'ethers';
-import { JsonRpcSigner } from '@ethersproject/providers';
+import { Contact, User, UserType } from '@types';
 
 import AccountService from './accountService';
 import ProviderService from './providerService';
@@ -34,7 +31,6 @@ describe('ContactList Service', () => {
   let meanApiService: jest.MockedObject<MeanApiService>;
   let contractService: jest.MockedObject<ContractService>;
   let contactListService: ContactListService;
-  const signerSendTransaction = jest.fn();
 
   beforeEach(() => {
     accountService = createMockInstance(MockedAccountService);
@@ -44,10 +40,6 @@ describe('ContactList Service', () => {
     contactListService = new ContactListService(accountService, providerService, meanApiService, contractService);
 
     accountService.getUser.mockReturnValue(userMock);
-    const mockedSigner = {
-      sendTransaction: signerSendTransaction.mockResolvedValue('sendTransaction'),
-    } as unknown as jest.Mocked<JsonRpcSigner>;
-    providerService.getSigner.mockResolvedValue(mockedSigner);
   });
 
   afterEach(() => {
@@ -164,111 +156,6 @@ describe('ContactList Service', () => {
         userMock.id
       );
       expect(contactListService.getContacts()).toEqual([contactMock]);
-    });
-  });
-
-  describe('transferTokenToContact', () => {
-    const erc20TokenMock = emptyTokenWithAddress('token', TokenType.ERC20_TOKEN);
-    const nativeTokenMock = emptyTokenWithAddress('nativeToken', TokenType.BASE);
-    const erc721TokenMock = emptyTokenWithAddress('nftToken', TokenType.ERC721_TOKEN);
-
-    test('it should not proceed if amount is not greater than zero', async () => {
-      try {
-        await contactListService.transferTokenToContact({
-          from: 'from',
-          contact: contactMock,
-          token: nativeTokenMock,
-          amount: BigNumber.from(0),
-        });
-        expect(1).toEqual(2);
-      } catch (e) {
-        // eslint-disable-next-line jest/no-conditional-expect
-        expect(e).toEqual(Error('Amount must be greater than zero'));
-      }
-    });
-    test("it should transfer tokens using the ERC20 interface if it's an ERC20 token", async () => {
-      const transferFn = jest.fn().mockResolvedValue('transfer');
-      contractService.getERC20TokenInstance.mockResolvedValue({
-        transfer: transferFn,
-      } as unknown as ERC20Contract);
-
-      const txResponse = await contactListService.transferTokenToContact({
-        from: 'from',
-        contact: contactMock,
-        token: erc20TokenMock,
-        amount: BigNumber.from(10),
-      });
-
-      expect(txResponse).toEqual('transfer');
-      expect(transferFn).toHaveBeenCalledTimes(1);
-      expect(transferFn).toHaveBeenCalledWith(contactMock.address, BigNumber.from(10));
-    });
-    test('it should generate a transaction from the signer if the token is a protocol token', async () => {
-      const txResponse = await contactListService.transferTokenToContact({
-        from: 'from',
-        contact: contactMock,
-        token: nativeTokenMock,
-        amount: BigNumber.from(10),
-      });
-
-      expect(txResponse).toEqual('sendTransaction');
-      expect(signerSendTransaction).toHaveBeenCalledTimes(1);
-      expect(signerSendTransaction).toHaveBeenCalledWith({
-        from: 'from',
-        to: contactMock.address,
-        value: BigNumber.from(10),
-      });
-    });
-    test('it should not proceed if token is not of type Base or ERC20', async () => {
-      try {
-        await contactListService.transferTokenToContact({
-          from: 'from',
-          contact: contactMock,
-          token: erc721TokenMock,
-          amount: BigNumber.from(10),
-        });
-        expect(1).toEqual(2);
-      } catch (e) {
-        // eslint-disable-next-line jest/no-conditional-expect
-        expect(e).toEqual(Error('Token must be of type Base or ERC20'));
-      }
-    });
-  });
-
-  describe('transferNFTToContact', () => {
-    const erc20TokenMock = emptyTokenWithAddress('token', TokenType.ERC20_TOKEN);
-    const erc721TokenMock = emptyTokenWithAddress('nftToken', TokenType.ERC721_TOKEN);
-
-    test('it should not proceed if token is not of type ERC721', async () => {
-      try {
-        await contactListService.transferNFTToContact({
-          from: 'from',
-          contact: contactMock,
-          token: erc20TokenMock,
-          tokenId: BigNumber.from(1111),
-        });
-        expect(1).toEqual(2);
-      } catch (e) {
-        // eslint-disable-next-line jest/no-conditional-expect
-        expect(e).toEqual(Error('Token must be of type ERC721'));
-      }
-    });
-    test("it should transfer token using the ERC721 interface if it's an ERC721 token", async () => {
-      const transferFromFn = jest.fn().mockResolvedValue('transferFrom');
-      contractService.getERC721TokenInstance.mockResolvedValue({
-        transferFrom: transferFromFn,
-      } as unknown as ERC721Contract);
-
-      const txResponse = await contactListService.transferNFTToContact({
-        from: 'from',
-        contact: contactMock,
-        token: erc721TokenMock,
-        tokenId: BigNumber.from(1111),
-      });
-
-      expect(txResponse).toEqual('transferFrom');
-      expect(transferFromFn).toHaveBeenCalledTimes(1);
-      expect(transferFromFn).toHaveBeenCalledWith('from', contactMock.address, BigNumber.from(1111));
     });
   });
