+++ apps/root/src/types/contactList.ts
@@ -1,7 +1,5 @@
-import { IAccountService, Token } from '@types';
+import { IAccountService } from '@types';
 import ProviderService from '@services/providerService';
-import { BigNumber } from 'ethers';
-import { TransactionResponse } from '@ethersproject/providers';
 
 export type Contact = {
   address: string;
@@ -25,23 +23,5 @@ export type IContactListService = {
   removeContact(contact: Contact): Promise<void>;
   editContact(contact: Contact): Promise<void>;
   getContacts(): ContactList;
-  transferTokenToContact({
-    contact,
-    token,
-    amount,
-  }: {
-    contact: Contact;
-    token: Token;
-    amount: BigNumber;
-  }): Promise<TransactionResponse | undefined>;
-  transferNFTToContact({
-    contact,
-    token,
-    tokenId,
-  }: {
-    contact: Contact;
-    token: Token;
-    tokenId: BigNumber;
-  }): Promise<TransactionResponse | undefined>;
   setContacts(contacts: ContactList): void;
 };
